Convert output to be more human readable
